auth/fix: auth fails with sha plugins when longer passwords used
on debug we noticed that sometimes server is sending NULL
terminated authPluginData in handshake-response or authswitch-response.

the fix is take only first 20 bytes of authPluginData

below is what other similar libraries say:

second part of the password cipher [mininum 13 bytes],
where len=MAX(13, length of auth-plugin-data - 8)

The web documentation is ambiguous about the length. However,
according to mysql-5.7/sql/auth/sql_authentication.cc line 538,
the 13th byte is "\0 byte, terminating the second part of
a scramble". So the second part of the password cipher is
a NULL terminated string that's at least 13 bytes with the
last byte being NULL.

The official Python library uses the fixed length 12
which seems to work but technically could have a hidden bug.